Utility to return the first Signal unlatched.
This is how we join the split when we create a new stack. We usually
want to wait until the new stack has started and it is ready to serve
its purpose, or otherwise be notified of an error that has brought the
whole program down. This lets us wait on multiple signals and return on
the first one. If more than one signal is already unlatched when
`Signal.first` is called, the result will be the result of the first
unlatched Signal.