Remove unused `occupied` flag.
Not sure where it is used. Not sure why you would use it. The notify
side of the signal shouldn't want to know anything about listeners.
In practice I notify with an error to error-first callbacks to indicate
an abort of some sort. Often times there is no notification if the
Signal is used as a cancel or other sort of interrupt.

Although I'm certain I don't use this anywhere, I'm still going to do a
breaking-change version bump.